ref(worker): wip
