+++ worker/worker/utils/task_queue.py
@@ -1,4 +1,3 @@
-import json
 import typing as tp
 
 import pika
@@ -20,17 +19,3 @@ def get_pika_connection(host: tp.Optional[str] = None, port: tp.Optional[int] = 
     channel = connection.channel()
     channel.queue_declare(queue=settings.pika_queue, durable=True)
     return connection, channel
-
-
-def send_transcription_job_to_queue(
-    channel: pika.channel.Channel,
-    queue_name: str,
-    transcription_id: int,
-    user_id: int,
-) -> tp.Mapping[str, tp.Any]:
-    job_dict = {
-        "transcription_id": transcription_id,
-        "user_id": user_id,
-    }
-    channel.basic_publish(exchange="", routing_key=queue_name, body=json.dumps(job_dict))
-    return job_dict
